Refactor routing structure by modularizing routes for tutorials, comments, and community notes; remove legacy database server scripts and update tests to include community note model mocks.
