--- expressServer/startDbClient.PopulateSampleData.cmd
@@ -1,1 +1,0 @@
-mongo --port 3000 -u dbAdmin -p test --authenticationDatabase admin
